US-24 mostly ready to go.
Final bug has to do with identifying the latest user preference set. Once the DB is cleaned up, just need to verify the MW meshes well.

Aside from that, add to cart page modified accordingly. We show user's preferences as recommendations and allow them to be removed from the display on a button click. Also the page is dynamic so that it does not look weird when you do not have preferences or you remove all the recommendations from the page.